+++ toys-core/lib/toys/middleware/show_help.rb
@@ -53,6 +53,12 @@ module Toys
       DEFAULT_HELP_FLAGS = ["-?", "--help"].freeze
 
       ##
+      # Default usage flags
+      # @return [Array<String>]
+      #
+      DEFAULT_USAGE_FLAGS = ["--usage"].freeze
+
+      ##
       # Default recursive flags
       # @return [Array<String>]
       #
@@ -68,44 +74,52 @@ module Toys
       # Create a ShowHelp middleware.
       #
       # @param [Boolean,Array<String>,Proc] help_flags Specify flags to
-      #     activate help. The value may be any of the following:
+      #     display help. The value may be any of the following:
       #     *  An array of flags.
-      #     *  The `true` value to use {DEFAULT_HELP_FLAGS}. (Default)
-      #     *  The `false` value for no flags.
+      #     *  The `true` value to use {DEFAULT_HELP_FLAGS}.
+      #     *  The `false` value for no flags. (Default)
       #     *  A proc that takes a tool and returns any of the above.
-      # @param [Boolean] fallback_execution Cause the tool to display its own
-      #     help text if it does not otherwise have an executor. This is
-      #     mostly useful for groups, which have children but no executor.
-      #     Default is `true`.
+      # @param [Boolean,Array<String>,Proc] usage_flags Specify flags to
+      #     display usage. The value may be any of the following:
+      #     *  An array of flags.
+      #     *  The `true` value to use {DEFAULT_USAGE_FLAGS}.
+      #     *  The `false` value for no flags. (Default)
+      #     *  A proc that takes a tool and returns any of the above.
       # @param [Boolean,Array<String>,Proc] recursive_flags Specify flags
       #     to control recursive subtool search. The value may be any of the
       #     following:
       #     *  An array of flags.
-      #     *  The `true` value to use {DEFAULT_RECURSIVE_FLAGS}. (Default)
-      #     *  The `false` value for no flags.
+      #     *  The `true` value to use {DEFAULT_RECURSIVE_FLAGS}.
+      #     *  The `false` value for no flags. (Default)
       #     *  A proc that takes a tool and returns any of the above.
       # @param [Boolean,Array<String>,Proc] search_flags Specify flags
       #     to search subtools for a search term. The value may be any of
       #     the following:
       #     *  An array of flags.
-      #     *  The `true` value to use {DEFAULT_SEARCH_FLAGS}. (Default)
-      #     *  The `false` value for no flags.
+      #     *  The `true` value to use {DEFAULT_SEARCH_FLAGS}.
+      #     *  The `false` value for no flags. (Default)
       #     *  A proc that takes a tool and returns any of the above.
       # @param [Boolean] default_recursive Whether to search recursively for
       #     subtools by default. Default is `false`.
+      # @param [Boolean] fallback_execution Cause the tool to display its own
+      #     help text if it does not otherwise have an executor. This is
+      #     mostly useful for groups, which have children but no executor.
+      #     Default is `true`.
       # @param [IO] stream Output stream to write to. Default is stdout.
       # @param [Boolean,nil] styled_output Cause the tool to display help text
       #     with ansi styles. If `nil`, display styles if the output stream is
       #     a tty. Default is `nil`.
       #
       def initialize(help_flags: false,
-                     fallback_execution: false,
-                     recursive_flags: true,
-                     search_flags: true,
+                     usage_flags: false,
+                     recursive_flags: false,
+                     search_flags: false,
                      default_recursive: false,
+                     fallback_execution: false,
                      stream: $stdout,
                      styled_output: nil)
         @help_flags = help_flags
+        @usage_flags = usage_flags
         @fallback_execution = fallback_execution
         @recursive_flags = recursive_flags
         @search_flags = search_flags
@@ -116,14 +130,20 @@ module Toys
       ##
       # Configure flags and default data.
       #
-      def config(tool)
+      def config(tool, loader)
         help_flags = Middleware.resolve_flags_spec(@help_flags, tool,
                                                    DEFAULT_HELP_FLAGS)
         unless help_flags.empty?
-          desc = "Show help message"
-          tool.add_flag(:_help, *help_flags, desc: desc, only_unique: true)
+          desc = "Display help for this tool"
+          tool.add_flag(:_show_help, *help_flags, desc: desc, only_unique: true)
         end
-        if !help_flags.empty? || @fallback_execution
+        usage_flags = Middleware.resolve_flags_spec(@usage_flags, tool,
+                                                    DEFAULT_USAGE_FLAGS)
+        unless usage_flags.empty?
+          desc = "Display a brief usage string for this tool"
+          tool.add_flag(:_show_usage, *usage_flags, desc: desc, only_unique: true)
+        end
+        if (!help_flags.empty? || @fallback_execution) && loader.has_subtools?(tool.full_name)
           add_recursive_flags(tool)
           add_search_flags(tool)
         end
@@ -134,15 +154,16 @@ module Toys
       # Display help text if requested.
       #
       def execute(context)
-        execute_by_fallback = @fallback_execution && !context[Context::TOOL].includes_executor?
-        execute_by_flag = context[:_help]
-        if execute_by_fallback || execute_by_flag
+        if @fallback_execution && !context[Context::TOOL].includes_executor? || context[:_show_help]
           help_text = Utils::HelpText.from_context(context)
-          width = ::HighLine.new.output_cols
           str = help_text.help_string(recursive: context[:_recursive_subtools],
                                       search: context[:_search_subtools],
                                       show_path: context[Context::VERBOSITY] > 0,
-                                      wrap_width: width)
+                                      wrap_width: ::HighLine.new.output_cols)
+          @output.puts(str)
+        elsif context[:_show_usage]
+          help_text = Utils::HelpText.from_context(context)
+          str = help_text.usage_string(wrap_width: ::HighLine.new.output_cols)
           @output.puts(str)
         else
           yield